Remove transcrit file if it is more than 30 days old (close #1419) (#1451)
* Remove transcrit file if it is more than 30 days old (close #1419)

* Fix isort

* Apply yapf formatter

+++ src/sos/workflow_engines.py
@@ -8,12 +8,12 @@ import stat
 import subprocess
 import time
 
-
 from .eval import cfg_interpolate
-from .utils import env, textMD5, expand_time, format_duration
+from .utils import env, expand_time, format_duration, textMD5
 
 
 class WorkflowEngine:
+
     def __init__(self, agent):
         self.agent = agent
         self.config = agent.config
@@ -26,9 +26,11 @@ class WorkflowEngine:
         else:
             r_idx = r_idx[0]
         # find next option
-        r_next = [idx for idx, x in enumerate(argv[r_idx + 1 :]) if x.startswith("-")]
+        r_next = [
+            idx for idx, x in enumerate(argv[r_idx + 1:]) if x.startswith("-")
+        ]
         if r_next:
-            argv = argv[:r_idx] + argv[r_idx + 1 + r_next[0] :]
+            argv = argv[:r_idx] + argv[r_idx + 1 + r_next[0]:]
         else:
             argv = argv[:r_idx]
         return argv
@@ -39,8 +41,8 @@ class WorkflowEngine:
             self.template_args["command"] = self.command
             self.template_args["job_name"] = self.job_name
             self.job_text = (
-                cfg_interpolate(self.workflow_template, self.template_args) + "\n"
-            )
+                cfg_interpolate(self.workflow_template, self.template_args) +
+                "\n")
         except Exception as e:
             raise ValueError(
                 f"Failed to generate job file for the execution of workflow: {e}"
@@ -64,21 +66,21 @@ class WorkflowEngine:
 
     def execute_workflow(self, filename, command, **template_args):
         # there is no need to prepare workflow (e.g. copy over)
-        from .parser import SoS_Script
         from .__main__ import get_run_parser
+        from .parser import SoS_Script
 
         parser = get_run_parser(interactive=False, with_workflow=True)
         args, workflow_args = parser.parse_known_args(command[2:])
         script = SoS_Script(filename=filename)
-        workflow = script.workflow(args.workflow, use_default=not args.__targets__)
+        workflow = script.workflow(
+            args.workflow, use_default=not args.__targets__)
 
         # we send config files also to remote host, but change localhost
         import yaml
 
         # process -c configfile
         host_cfg_file = os.path.join(
-            os.path.expanduser("~"), ".sos", f"config_{self.alias}.yml"
-        )
+            os.path.expanduser("~"), ".sos", f"config_{self.alias}.yml")
         with open(host_cfg_file, "w") as cfg:
             remote_cfg = copy.deepcopy(env.sos_dict["CONFIG"])
             remote_cfg["localhost"] = self.alias
@@ -116,12 +118,12 @@ class WorkflowEngine:
             self.command[1] = self.filename
         else:
             raise ValueError(
-                f"Failed to generate remote execution command: {self.command}"
-            )
+                f"Failed to generate remote execution command: {self.command}")
         self.command = subprocess.list2cmdline(self.command)
         self.template_args = copy.deepcopy(self.config)
         self.template_args.update(template_args)
-        env.log_to_file("WORKFLOW", f"Execute command on remote host: {self.command}")
+        env.log_to_file("WORKFLOW",
+                        f"Execute command on remote host: {self.command}")
         return True
 
     def query_workflows(
@@ -147,8 +149,7 @@ class WorkflowEngine:
                     f"--age {age}" if age else "",
                     f'--tags {" ".join(tags)}' if tags else "",
                     f'--status {" ".join(status)}' if status else "",
-                )
-            )
+                ))
         except subprocess.CalledProcessError as e:
             if verbosity >= 3:
                 env.logger.warning(
@@ -169,16 +170,18 @@ class WorkflowEngine:
             env.logger.debug(f'"{cmd}" executed with response "{ret}"')
         except subprocess.CalledProcessError:
             env.logger.error(
-                "Failed to kill all workflows"
-                if all_workflows
-                else f'Failed to kill workflows {" ".join(workflows)}'
-            )
+                "Failed to kill all workflows" if all_workflows else
+                f'Failed to kill workflows {" ".join(workflows)}')
             return ""
         return ret
 
-    def purge_workflows(
-        self, workflows, purge_all=False, age=None, status=None, tags=None, verbosity=2
-    ):
+    def purge_workflows(self,
+                        workflows,
+                        purge_all=False,
+                        age=None,
+                        status=None,
+                        tags=None,
+                        verbosity=2):
         try:
             return self.agent.check_output(
                 "{} purge {} {} {} {} {} -v {}".format(
@@ -186,23 +189,23 @@ class WorkflowEngine:
                     " ".join(workflows),
                     "--all" if purge_all else "",
                     f"--age {age}" if age is not None else "",
-                    f'--status {" ".join(status)}' if status is not None else "",
+                    f'--status {" ".join(status)}'
+                    if status is not None else "",
                     f'--tags {" ".join(tags)}' if tags is not None else "",
                     verbosity,
-                )
-            )
+                ))
         except subprocess.CalledProcessError:
             env.logger.error(f"Failed to purge workflows {workflows}")
             return ""
 
 
 class BackgroundProcess_WorkflowEngine(WorkflowEngine):
+
     def __init__(self, agent):
         super(BackgroundProcess_WorkflowEngine, self).__init__(agent)
         if "workflow_template" in self.config:
             self.workflow_template = self.config["workflow_template"].replace(
-                "\r\n", "\n"
-            )
+                "\r\n", "\n")
         else:
             self.workflow_template = None
 
@@ -211,11 +214,10 @@ class BackgroundProcess_WorkflowEngine(WorkflowEngine):
         # calling super execute_workflow would set cleaned versions
         # of self.filename, self.command, and self.template_args
         if not super(BackgroundProcess_WorkflowEngine, self).execute_workflow(
-            filename, command, **template_args
-        ):
+                filename, command, **template_args):
             env.log_to_file(
-                "WORKFLOW", f'Failed to prepare workflow with command "{command}"'
-            )
+                "WORKFLOW",
+                f'Failed to prepare workflow with command "{command}"')
             return False
 
         if self.workflow_template:
@@ -245,8 +247,8 @@ class BackgroundProcess_WorkflowEngine(WorkflowEngine):
 
             cmd = f"bash ~/.sos/workflows/{os.path.basename(self.job_file)}"
             env.log_to_file(
-                "WORKFLOW", f'Execute "{self.command}" with script {self.job_text}'
-            )
+                "WORKFLOW",
+                f'Execute "{self.command}" with script {self.job_text}')
             self.agent.check_call(cmd, under_workdir=True)
         except Exception as e:
             raise RuntimeError(
@@ -256,11 +258,13 @@ class BackgroundProcess_WorkflowEngine(WorkflowEngine):
             try:
                 os.remove(self.job_file)
             except Exception as e:
-                env.logger.debug(f"Failed to remove temporary workflow file: {e}")
+                env.logger.debug(
+                    f"Failed to remove temporary workflow file: {e}")
         return True
 
 
 class WorkflowPulse:
+
     def __init__(self, workflow_id):
         self.id = workflow_id
         self._tags = None
@@ -269,8 +273,7 @@ class WorkflowPulse:
         self._start_time = None
         self._complete_time = None
         self.pulse_file = os.path.join(
-            os.path.expanduser("~"), ".sos", "workflows", self.id + ".pulse"
-        )
+            os.path.expanduser("~"), ".sos", "workflows", self.id + ".pulse")
 
     def exists(self):
         return os.path.isfile(self.pulse_file)
@@ -282,25 +285,22 @@ class WorkflowPulse:
 
     @property
     def created(self):
-        return (
-            "Created " + format_duration(time.time() - self._start_time, True) + " ago"
-            if self._start_time
-            else ""
-        )
+        return ("Created " +
+                format_duration(time.time() - self._start_time, True) +
+                " ago" if self._start_time else "")
 
     @property
     def started(self):
-        return (
-            "Started " + format_duration(time.time() - self._start_time, True) + " ago"
-            if self._start_time
-            else ""
-        )
+        return ("Started " +
+                format_duration(time.time() - self._start_time, True) +
+                " ago" if self._start_time else "")
 
     @property
     def duration(self):
         if not self._complete_time:
             return ""
-        return "Ran for " + format_duration(int(self._complete_time - self._start_time))
+        return "Ran for " + format_duration(
+            int(self._complete_time - self._start_time))
 
     @property
     def tags(self):
@@ -319,8 +319,7 @@ class WorkflowPulse:
             return self._content[ext]
 
         res_file = os.path.join(
-            os.path.expanduser("~"), ".sos", "workflows", self.id + ext
-        )
+            os.path.expanduser("~"), ".sos", "workflows", self.id + ext)
         content = ""
         if os.path.isfile(res_file):
             try:
@@ -385,7 +384,8 @@ class WorkflowPulse:
                     self._tags = fields[2].strip()
         if not os.stat(self.pulse_file).st_mode & stat.S_IWUSR:
             self._status = 'aborted'
-        if last_active_time is None or time.time() - float(last_active_time) > 120:
+        if last_active_time is None or time.time() - float(
+                last_active_time) > 120:
             self._status = 'failed'
 
 
@@ -404,8 +404,8 @@ def print_workflow_status(
     all_workflows: List = []
     if check_all:
         workflows = glob.glob(
-            os.path.join(os.path.expanduser("~"), ".sos", "workflows", "*.pulse")
-        )
+            os.path.join(
+                os.path.expanduser("~"), ".sos", "workflows", "*.pulse"))
         all_workflows = [
             (os.path.basename(x)[:-6], os.path.getmtime(x)) for x in workflows
         ]
@@ -415,12 +415,10 @@ def print_workflow_status(
         for t in workflows:
             matched_names = glob.glob(
                 os.path.join(
-                    os.path.expanduser("~"), ".sos", "workflows", f"{t}*.pulse"
-                )
-            )
-            matched = [
-                (os.path.basename(x)[:-6], os.path.getmtime(x)) for x in matched_names
-            ]
+                    os.path.expanduser("~"), ".sos", "workflows",
+                    f"{t}*.pulse"))
+            matched = [(os.path.basename(x)[:-6], os.path.getmtime(x))
+                       for x in matched_names]
             if not matched:
                 all_workflows.append((t, None))
             else:
@@ -429,20 +427,21 @@ def print_workflow_status(
     if age is not None:
         age = expand_time(age, default_unit="d")
         if age > 0:
-            all_workflows = [x for x in all_workflows if time.time() - x[1] >= age]
+            all_workflows = [
+                x for x in all_workflows if time.time() - x[1] >= age
+            ]
         else:
-            all_workflows = [x for x in all_workflows if time.time() - x[1] <= -age]
+            all_workflows = [
+                x for x in all_workflows if time.time() - x[1] <= -age
+            ]
 
     all_workflows = sorted(
-        list(set(all_workflows)), key=lambda x: 0 if x[1] is None else x[1]
-    )
+        list(set(all_workflows)), key=lambda x: 0 if x[1] is None else x[1])
 
     if tags:
         all_workflows = [
-            x
-            for x in all_workflows
-            if WorkflowPulse(x[0]).exists()
-            and any(y in tags for y in WorkflowPulse(x[0]).tags.split())
+            x for x in all_workflows if WorkflowPulse(x[0]).exists() and any(
+                y in tags for y in WorkflowPulse(x[0]).tags.split())
         ]
 
     if not all_workflows:
@@ -452,9 +451,9 @@ def print_workflow_status(
     #
     # automatically remove non-running workflows that are more than 30 days old
     to_be_removed = [
-        t
-        for s, (t, d) in zip(workflow_info, all_workflows)
-        if d is not None and time.time() - d > 30 * 24 * 60 * 60 and s != "running"
+        t for s, (t, d) in zip(workflow_info, all_workflows)
+        if d is not None and time.time() - d > 30 * 24 * 60 *
+        60 and s != "running"
     ]
 
     if status:
@@ -525,12 +524,14 @@ def kill_workflow(workflow):
     if status == "completed":
         return "completed"
     with open(
-        os.path.join(
-            os.path.expanduser("~"), ".sos", "workflows", workflow + ".soserr"
-        ),
-        "a",
+            os.path.join(
+                os.path.expanduser("~"), ".sos", "workflows",
+                workflow + ".soserr"),
+            "a",
     ) as err:
-        err.write(f"Workflow {workflow} killed by sos kill command or workflow engine.")
+        err.write(
+            f"Workflow {workflow} killed by sos kill command or workflow engine."
+        )
 
     wp.mark_killed()
     return "aborted"
@@ -541,17 +542,16 @@ def kill_workflows(workflows, tags=None):
 
     if not workflows:
         workflows = glob.glob(
-            os.path.join(os.path.expanduser("~"), ".sos", "workflows", "*.pulse")
-        )
+            os.path.join(
+                os.path.expanduser("~"), ".sos", "workflows", "*.pulse"))
         all_workflows = [os.path.basename(x)[:-6] for x in workflows]
     else:
         all_workflows = []
         for t in workflows:
             matched = glob.glob(
                 os.path.join(
-                    os.path.expanduser("~"), ".sos", "workflows", f"{t}*.pulse"
-                )
-            )
+                    os.path.expanduser("~"), ".sos", "workflows",
+                    f"{t}*.pulse"))
             matched = [os.path.basename(x)[:-6] for x in matched]
             if not matched:
                 env.logger.warning(f"{t} does not match any existing workflow")
@@ -559,9 +559,8 @@ def kill_workflows(workflows, tags=None):
                 all_workflows.extend(matched)
     if tags:
         all_workflows = [
-            x
-            for x in all_workflows
-            if any(x in tags for x in WorkflowPulse(x).tags.split())
+            x for x in all_workflows if any(
+                x in tags for x in WorkflowPulse(x).tags.split())
         ]
 
     if not all_workflows:
@@ -574,9 +573,12 @@ def kill_workflows(workflows, tags=None):
         print(f"{wf}\t{ret}")
 
 
-def purge_workflows(
-    workflows, purge_all=None, age=None, status=None, tags=None, verbosity=2
-):
+def purge_workflows(workflows,
+                    purge_all=None,
+                    age=None,
+                    status=None,
+                    tags=None,
+                    verbosity=2):
     import glob
 
     if workflows:
@@ -584,17 +586,18 @@ def purge_workflows(
         for t in workflows:
             matched = glob.glob(
                 os.path.join(
-                    os.path.expanduser("~"), ".sos", "workflows", f"{t}*.pulse"
-                )
-            )
-            matched = [(os.path.basename(x)[:-6], os.path.getmtime(x)) for x in matched]
+                    os.path.expanduser("~"), ".sos", "workflows",
+                    f"{t}*.pulse"))
+            matched = [
+                (os.path.basename(x)[:-6], os.path.getmtime(x)) for x in matched
+            ]
             if not matched:
                 print(f"{t}\tmissing")
             all_workflows.extend(matched)
     elif purge_all or age or status or tags:
         workflows = glob.glob(
-            os.path.join(os.path.expanduser("~"), ".sos", "workflows", "*.pulse")
-        )
+            os.path.join(
+                os.path.expanduser("~"), ".sos", "workflows", "*.pulse"))
         all_workflows = [
             (os.path.basename(x)[:-6], os.path.getmtime(x)) for x in workflows
         ]
@@ -606,9 +609,13 @@ def purge_workflows(
     if age is not None:
         age = expand_time(age, default_unit="d")
         if age > 0:
-            all_workflows = [x for x in all_workflows if time.time() - x[1] >= age]
+            all_workflows = [
+                x for x in all_workflows if time.time() - x[1] >= age
+            ]
         else:
-            all_workflows = [x for x in all_workflows if time.time() - x[1] <= -age]
+            all_workflows = [
+                x for x in all_workflows if time.time() - x[1] <= -age
+            ]
 
     if status:
         all_workflows = [
@@ -617,9 +624,8 @@ def purge_workflows(
 
     if tags:
         all_workflows = [
-            x
-            for x in all_workflows
-            if any(x in tags for x in WorkflowPulse(x[0]).tags.split())
+            x for x in all_workflows if any(
+                x in tags for x in WorkflowPulse(x[0]).tags.split())
         ]
     #
     # remoe all workflow files
@@ -631,8 +637,7 @@ def purge_workflows(
 
         to_be_removed = defaultdict(list)
         for dirname, _, filelist in os.walk(
-            os.path.join(os.path.expanduser("~"), ".sos", "workflows")
-        ):
+                os.path.join(os.path.expanduser("~"), ".sos", "workflows")):
             for f in filelist:
                 ID = os.path.basename(f).split(".", 1)[0]
                 if ID in all_workflows:
@@ -644,18 +649,15 @@ def purge_workflows(
             for f in to_be_removed[workflow]:
                 try:
                     if verbosity > 3:
-                        if (
-                            "WORKFLOW" in env.config["SOS_DEBUG"]
-                            or "ALL" in env.config["SOS_DEBUG"]
-                        ):
+                        if ("WORKFLOW" in env.config["SOS_DEBUG"] or
+                                "ALL" in env.config["SOS_DEBUG"]):
                             env.log_to_file("TASK", f"Remove {f}")
                     os.remove(f)
                 except Exception as e:
                     removed = False
                     if verbosity > 0:
                         env.logger.warning(
-                            f"Failed to purge workflow {workflow[0]}: {e}"
-                        )
+                            f"Failed to purge workflow {workflow[0]}: {e}")
             status_cache.pop(workflow, None)
             if removed and verbosity > 1:
                 print(f"{workflow}\tpurged")
@@ -663,8 +665,7 @@ def purge_workflows(
         env.logger.debug("No matching workflows to purge")
     if purge_all and age is None and status is None and tags is None:
         matched = glob.glob(
-            os.path.join(os.path.expanduser("~"), ".sos", "workflows", "*")
-        )
+            os.path.join(os.path.expanduser("~"), ".sos", "workflows", "*"))
         count = 0
         for f in matched:
             if os.path.isdir(f):
